chore: extract webidl from fetch folder
