+++ lib/web/cache/cachestorage.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { Cache } = require('./cache')
-const { webidl } = require('../fetch/webidl')
+const { webidl } = require('../webidl')
 const { kEnumerableProperty } = require('../../core/util')
 const { kConstruct } = require('../../core/symbols')
