+++ lib/web/fetch/body.js
@@ -10,7 +10,7 @@ const {
   utf8DecodeBytes
 } = require('./util')
 const { FormData, setFormDataState } = require('./formdata')
-const { webidl } = require('./webidl')
+const { webidl } = require('../webidl')
 const { Blob } = require('node:buffer')
 const assert = require('node:assert')
 const { isErrored, isDisturbed } = require('node:stream')
