+++ lib/web/fetch/formdata-parser.js
@@ -4,7 +4,7 @@ const { isUSVString, bufferToLowerCasedHeaderName } = require('../../core/util')
 const { utf8DecodeBytes } = require('./util')
 const { HTTP_TOKEN_CODEPOINTS, isomorphicDecode } = require('./data-url')
 const { makeEntry } = require('./formdata')
-const { webidl } = require('./webidl')
+const { webidl } = require('../webidl')
 const assert = require('node:assert')
 const { File: NodeFile } = require('node:buffer')
