+++ lib/web/fetch/response.js
@@ -18,7 +18,7 @@ const {
   redirectStatusSet,
   nullBodyStatus
 } = require('./constants')
-const { webidl } = require('./webidl')
+const { webidl } = require('../webidl')
 const { URLSerializer } = require('./data-url')
 const { kConstruct } = require('../../core/symbols')
 const assert = require('node:assert')
