+++ lib/web/fetch/util.js
@@ -9,7 +9,7 @@ const { performance } = require('node:perf_hooks')
 const { ReadableStreamFrom, isValidHTTPToken, normalizedMethodRecordsBase } = require('../../core/util')
 const assert = require('node:assert')
 const { isUint8Array } = require('node:util/types')
-const { webidl } = require('./webidl')
+const { webidl } = require('../webidl')
 
 let supportedHashes = []
